requirements.txt
attrs==22.2.0
jsonschema==4.17.3
mysql-connector-python==8.0.32
ocpp==0.17.0
protobuf==3.20.3
pyrsistent==0.19.3
python-dotenv==0.21.1
websockets==10.4
